rename src/parser/tbc to src/parser/classic
aligns directory structure with analysis directory
