+++ src/analysis/classic/druid/CombatLogParser.ts
@@ -3,8 +3,8 @@ import AbilityTracker from 'parser/shared/modules/AbilityTracker';
 import ManaLevelChart from 'parser/shared/modules/resources/mana/ManaLevelChart';
 import ManaUsageChart from 'parser/shared/modules/resources/mana/ManaUsageChart';
 import SpellManaCost from 'parser/shared/modules/SpellManaCost';
-import BaseCombatLogParser from 'parser/tbc/CombatLogParser';
-import lowRankSpellsSuggestion from 'parser/tbc/suggestions/lowRankSpells';
+import BaseCombatLogParser from 'parser/classic/CombatLogParser';
+import lowRankSpellsSuggestion from 'parser/classic/suggestions/lowRankSpells';
 
 import lowRankSpells from './lowRankSpells';
 import Abilities from './modules/Abilities';
