+++ src/analysis/classic/mage/CombatLogParser.ts
@@ -1,6 +1,6 @@
 import Channeling from 'parser/shared/normalizers/Channeling';
-import BaseCombatLogParser from 'parser/tbc/CombatLogParser';
-import lowRankSpellsSuggestion from 'parser/tbc/suggestions/lowRankSpells';
+import BaseCombatLogParser from 'parser/classic/CombatLogParser';
+import lowRankSpellsSuggestion from 'parser/classic/suggestions/lowRankSpells';
 
 import lowRankSpells from './lowRankSpells';
 import Abilities from './modules/Abilities';
