+++ src/analysis/classic/paladin/CombatLogParser.ts
@@ -3,9 +3,9 @@ import AbilityTracker from 'parser/shared/modules/AbilityTracker';
 import ManaLevelChart from 'parser/shared/modules/resources/mana/ManaLevelChart';
 import ManaUsageChart from 'parser/shared/modules/resources/mana/ManaUsageChart';
 import SpellManaCost from 'parser/shared/modules/SpellManaCost';
-import BaseCombatLogParser from 'parser/tbc/CombatLogParser';
-import PreparationRuleAnalyzer from 'parser/tbc/modules/features/Checklist/PreparationRuleAnalyzer';
-import lowRankSpellsSuggestion from 'parser/tbc/suggestions/lowRankSpells';
+import BaseCombatLogParser from 'parser/classic/CombatLogParser';
+import PreparationRuleAnalyzer from 'parser/classic/modules/features/Checklist/PreparationRuleAnalyzer';
+import lowRankSpellsSuggestion from 'parser/classic/suggestions/lowRankSpells';
 
 import lowRankSpells, { whitelist } from './lowRankSpells';
 import Abilities from './modules/Abilities';
