+++ src/analysis/classic/priest/modules/checklist/Component.tsx
@@ -9,7 +9,7 @@ import {
 import GenericCastEfficiencyRequirement from 'parser/shared/modules/features/Checklist/GenericCastEfficiencyRequirement';
 import Requirement from 'parser/shared/modules/features/Checklist/Requirement';
 import Rule from 'parser/shared/modules/features/Checklist/Rule';
-import PreparationRule from 'parser/tbc/modules/features/Checklist/PreparationRule';
+import PreparationRule from 'parser/classic/modules/features/Checklist/PreparationRule';
 
 import * as SPELLS from '../../SPELLS';
