+++ src/analysis/classic/rogue/modules/checklist/Module.tsx
@@ -1,7 +1,7 @@
 import CastEfficiency from 'parser/shared/modules/CastEfficiency';
 import Combatants from 'parser/shared/modules/Combatants';
 import BaseChecklist from 'parser/shared/modules/features/Checklist/Module';
-import PreparationRuleAnalyzer from 'parser/tbc/modules/features/Checklist/PreparationRuleAnalyzer';
+import PreparationRuleAnalyzer from 'parser/classic/modules/features/Checklist/PreparationRuleAnalyzer';
 
 import AlwaysBeCasting from '../features/AlwaysBeCasting';
 import Component from './Component';
