+++ src/analysis/classic/warrior/CombatLogParser.ts
@@ -1,5 +1,5 @@
-import BaseCombatLogParser from 'parser/tbc/CombatLogParser';
-import lowRankSpellsSuggestion from 'parser/tbc/suggestions/lowRankSpells';
+import BaseCombatLogParser from 'parser/classic/CombatLogParser';
+import lowRankSpellsSuggestion from 'parser/classic/suggestions/lowRankSpells';
 
 import lowRankSpells, { whitelist } from './lowRankSpells';
 import Abilities from './modules/Abilities';
