+++ src/parser/classic/modules/ManaValues.tsx
