+++ src/parser/classic/modules/items/CombatPotionChecker.tsx
