+++ src/parser/classic/modules/items/FoodChecker.ts
