+++ src/parser/classic/modules/items/InFightConsumablesChecker.tsx
