+++ src/parser/classic/modules/items/WeaponEnhancementChecker.tsx
