+++ src/parser/classic/statistic/ManaGained.tsx
