+++ src/parser/classic/suggestions/lowRankSpells.test.ts
